# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Omni - Documentação da API",
        "description": "Documentação de todos os endpoints da API do sistema Omni.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/collections": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "Lista todas as coleções",
                "operationId": "8e76c986c49f998da6ed83b146f38072",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Collection"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Collections"
                ],
                "summary": "Cria uma nova coleção",
                "operationId": "3239fa1a71f0eb35ff0b2018c2f58ce5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCollectionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/collections/{id}": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "Busca uma coleção pelo ID",
                "operationId": "7e5f174d68d28530f5ed08ef7ab7b9e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Coleção não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Collections"
                ],
                "summary": "Atualiza uma coleção",
                "operationId": "2c44a856324b0d24a5712804b8cf67df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCollectionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Coleção não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Collections"
                ],
                "summary": "Deleta uma coleção",
                "operationId": "2cc70dd529750ec0db88244ab51b3b1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletado com sucesso"
                    },
                    "404": {
                        "description": "Coleção não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCollectionRequest": {
                "title": "Store Collection Request",
                "required": [
                    "name",
                    "slug",
                    "owner_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Leads do Site"
                    },
                    "slug": {
                        "type": "string",
                        "example": "leads-site"
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "visibility": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private",
                            "unlisted"
                        ],
                        "example": "private"
                    },
                    "agency_access": {
                        "type": "boolean",
                        "example": true
                    },
                    "icon": {
                        "description": "Ícone associado à coleção",
                        "type": "string",
                        "example": "fa-users",
                        "nullable": true
                    },
                    "description": {
                        "description": "Descrição detalhada sobre o propósito da coleção",
                        "type": "string",
                        "nullable": true
                    },
                    "schema_json": {
                        "type": "object",
                        "example": {
                            "fields": [
                                {
                                    "name": "email",
                                    "type": "email"
                                },
                                {
                                    "name": "telefone",
                                    "type": "tel"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "UpdateCollectionRequest": {
                "title": "Update Collection Request",
                "description": "Todos os campos são opcionais. Envie apenas os que deseja alterar.",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Leads Qualificados do Site"
                    },
                    "slug": {
                        "type": "string",
                        "example": "leads-qualificados-site"
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "visibility": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private",
                            "unlisted"
                        ],
                        "example": "private"
                    },
                    "agency_access": {
                        "type": "boolean",
                        "example": false
                    },
                    "icon": {
                        "type": "string",
                        "example": "fa-star",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Coleção atualizada para leads qualificados.",
                        "nullable": true
                    },
                    "schema_json": {
                        "type": "object",
                        "example": {
                            "fields": [
                                {
                                    "name": "email",
                                    "type": "email"
                                },
                                {
                                    "name": "status",
                                    "type": "string"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "Collection": {
                "title": "Collection",
                "description": "Modelo de Coleção de Dados customizável",
                "properties": {
                    "id": {
                        "description": "ID único da coleção",
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "description": "Nome da coleção",
                        "type": "string",
                        "example": "Leads de Marketing"
                    },
                    "slug": {
                        "description": "URL amigável da coleção (geralmente gerado a partir do nome)",
                        "type": "string",
                        "example": "leads-de-marketing"
                    },
                    "client_id": {
                        "description": "ID do cliente ao qual esta coleção está associada (opcional)",
                        "type": "integer",
                        "nullable": true
                    },
                    "owner_id": {
                        "description": "ID do usuário que é dono da coleção",
                        "type": "integer",
                        "example": 1
                    },
                    "visibility": {
                        "description": "Define quem pode ver a coleção",
                        "type": "string",
                        "enum": [
                            "public",
                            "private",
                            "unlisted"
                        ],
                        "example": "public"
                    },
                    "agency_access": {
                        "description": "Indica se a agência tem acesso a esta coleção",
                        "type": "boolean",
                        "example": true
                    },
                    "icon": {
                        "description": "Ícone associado à coleção (ex: nome do ícone de uma biblioteca ou URL)",
                        "type": "string",
                        "example": "fa-users",
                        "nullable": true
                    },
                    "description": {
                        "description": "Descrição detalhada sobre o propósito da coleção",
                        "type": "string",
                        "nullable": true
                    },
                    "schema_json": {
                        "description": "Estrutura JSON que define os campos customizados da coleção",
                        "type": "object",
                        "example": {
                            "fields": [
                                {
                                    "name": "nome_completo",
                                    "type": "string",
                                    "required": true
                                },
                                {
                                    "name": "email",
                                    "type": "email",
                                    "required": true
                                }
                            ]
                        }
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}